Generate <classname>.cfg/userconfig.cfg files at runtime if they don't exist  * Closes #87 in conjunction with fortressforever/fortressforever-scripts#3
