+++ cl_dll/cdll_client_int.cpp
@@ -91,6 +91,11 @@
 #include <cl_dll/iviewport.h>
 // END: Added by Mulchman for team menu at level start up
 
+// <-- FF
+#include "ff_cdll_client_int.h"
+extern CFFClient gFFClient;
+// --> FF
+
 // memdbgon must be the last include file in a .cpp file!!!
 #include "tier0/memdbgon.h"
 
@@ -405,111 +410,15 @@ void DisplayBoneSetupEnts()
 #endif
 }
 
-
-//-----------------------------------------------------------------------------
-// Purpose: engine to client .dll interface
-//-----------------------------------------------------------------------------
-class CHLClient : public IBaseClientDLL
-{
-public:
-	CHLClient();
-
-	virtual int						Init( CreateInterfaceFn appSystemFactory, CreateInterfaceFn physicsFactory, CGlobalVarsBase *pGlobals );
-
-	virtual void					Shutdown( void );
-
-	virtual void					LevelInitPreEntity( const char *pMapName );
-	virtual void					LevelInitPostEntity();
-	virtual void					LevelShutdown( void );
-
-	virtual ClientClass				*GetAllClasses( void );
-
-	virtual int						HudVidInit( void );
-	virtual void					HudProcessInput( bool bActive );
-	virtual void					HudUpdate( bool bActive );
-	virtual void					HudReset( void );
-	virtual void					HudText( const char * message );
-
-	// Mouse Input Interfaces
-	virtual void					IN_ActivateMouse( void );
-	virtual void					IN_DeactivateMouse( void );
-	virtual void					IN_MouseEvent( int mstate, bool down );
-	virtual void					IN_Accumulate( void );
-	virtual void					IN_ClearStates( void );
-	virtual bool					IN_IsKeyDown( const char *name, bool& isdown );
-	// Raw signal
-	virtual int						IN_KeyEvent( int eventcode, int keynum, const char *pszCurrentBinding );
-	// Create movement command
-	virtual void					CreateMove ( int sequence_number, float input_sample_frametime, bool active );
-	virtual void					ExtraMouseSample( float frametime, bool active );
-	virtual bool					WriteUsercmdDeltaToBuffer( bf_write *buf, int from, int to, bool isnewcommand );	
-	virtual void					EncodeUserCmdToBuffer( bf_write& buf, int slot );
-	virtual void					DecodeUserCmdFromBuffer( bf_read& buf, int slot );
-
-
-	virtual void					View_Render( vrect_t *rect );
-	virtual void					RenderView( const CViewSetup &view, int nClearFlags, bool drawViewmodel );
-	virtual void					View_Fade( ScreenFade_t *pSF );
-	
-	virtual void					SetCrosshairAngle( const QAngle& angle );
-
-	virtual void					InitSprite( CEngineSprite *pSprite, const char *loadname );
-	virtual void					ShutdownSprite( CEngineSprite *pSprite );
-
-	virtual int						GetSpriteSize( void ) const;
-
-	virtual void					VoiceStatus( int entindex, qboolean bTalking );
-
-	virtual void					InstallStringTableCallback( const char *tableName );
-
-	virtual void					FrameStageNotify( ClientFrameStage_t curStage );
-
-	virtual bool					DispatchUserMessage( int msg_type, bf_read &msg_data );
-
-	// Save/restore system hooks
-	virtual CSaveRestoreData  *SaveInit( int size );
-	virtual void			SaveWriteFields( CSaveRestoreData *, const char *, void *, datamap_t *, typedescription_t *, int );
-	virtual void			SaveReadFields( CSaveRestoreData *, const char *, void *, datamap_t *, typedescription_t *, int );
-	virtual void			PreSave( CSaveRestoreData * );
-	virtual void			Save( CSaveRestoreData * );
-	virtual void			WriteSaveHeaders( CSaveRestoreData * );
-	virtual void			ReadRestoreHeaders( CSaveRestoreData * );
-	virtual void			Restore( CSaveRestoreData *, bool );
-	virtual void			DispatchOnRestore();
-	virtual void			WriteSaveGameScreenshot( const char *pFilename );
-
-	// Given a list of "S(wavname) S(wavname2)" tokens, look up the localized text and emit
-	//  the appropriate close caption if running with closecaption = 1
-	virtual void			EmitSentenceCloseCaption( char const *tokenstream );
-	virtual void			EmitCloseCaption( char const *captionname, float duration );
-
-	virtual CStandardRecvProxies* GetStandardRecvProxies();
-
-	virtual bool			CanRecordDemo( char *errorMsg, int length ) const;
-
-	// save game screenshot writing
-	virtual void			WriteSaveGameScreenshotOfSize( const char *pFilename, int width, int height );
-
-	// See RenderViewInfo_t
-	virtual void			RenderViewEx( const CViewSetup &view, int nClearFlags, int whatToDraw );
-
-	// Gets the location of the player viewpoint
-	virtual bool			GetPlayerView( CViewSetup &playerView );
-
-public:
-	void PrecacheMaterial( const char *pMaterialName );
-
-private:
-	void UncacheAllMaterials( );
-
-	CUtlVector< IMaterial * > m_CachedMaterials;
-};
-
+/*
+// Class definition moved to cdll_client_int.h, so that we can inherit from it
+// FF sets its own version of these variables
 
 CHLClient gHLClient;
 IBaseClientDLL *clientdll = &gHLClient;
 
 EXPOSE_SINGLE_INTERFACE_GLOBALVAR( CHLClient, IBaseClientDLL, CLIENT_DLL_INTERFACE_VERSION, gHLClient );
+*/
 
 
 //-----------------------------------------------------------------------------
@@ -517,7 +426,7 @@ EXPOSE_SINGLE_INTERFACE_GLOBALVAR( CHLClient, IBaseClientDLL, CLIENT_DLL_INTERFA
 //-----------------------------------------------------------------------------
 void PrecacheMaterial( const char *pMaterialName )
 {
-	gHLClient.PrecacheMaterial( pMaterialName );
+	gFFClient.PrecacheMaterial( pMaterialName );
 }
 
 //-----------------------------------------------------------------------------
@@ -1301,7 +1210,7 @@ void CHLClient::VoiceStatus( int entindex, qboolean bTalking )
 void OnMaterialStringTableChanged( void *object, INetworkStringTable *stringTable, int stringNumber, const char *newString, void const *newData )
 {
 	// Make sure this puppy is precached
-	gHLClient.PrecacheMaterial( newString );
+	gFFClient.PrecacheMaterial( newString );
 
 	RequestCacheUsedMaterials();
 }
